requirements.txt
azure-devops==7.1.0b4
nltk==3.8.1
Flask==3.0.3
